Server/app/api/requirements.txt
flask>=2.1.3
pandas==1.4.3
scikit-learn==1.1.1
gevent==22.10.2
Pyrebase4==4.5.0
# opencv-python-headless==4.5.1.48
opencv-python==4.5.5.64
mediapipe==0.8.11